fix bug

+++ src/main/scala/com/cpc/spark/common/CpcPartitioner.scala
@@ -5,7 +5,7 @@ import org.apache.spark.Partitioner
 import scala.util.hashing.MurmurHash3
 
 class CpcPartitioner(num: Int) extends Partitioner {
-  override def numPartitions: Long = num.toLong
+  override def numPartitions: Int = num
 
   override def getPartition(key: Any): Int = {
     val k = key.asInstanceOf[(String, Int)]
